+++ packages/kit/src/runtime/server/page/render.js
@@ -14,7 +14,7 @@ import { create_async_iterator } from '../../../utils/streaming.js';
 import { SVELTE_KIT_ASSETS } from '../../../constants.js';
 import { SCHEME } from '../../../utils/url.js';
 import { create_server_routing_response, generate_route_object } from './server_routing.js';
-import { add_resolution_prefix } from '../../pathname.js';
+import { add_resolution_suffix } from '../../pathname.js';
 
 // TODO rename this function/module
 
@@ -116,11 +116,6 @@ export async function render_response({
 	}
 
 	if (page_config.ssr) {
-		if (__SVELTEKIT_DEV__ && !branch.at(-1)?.node.component) {
-			// Can only be the leaf, layouts have a fallback component generated
-			throw new Error(`Missing +page.svelte component for route ${event.route.id}`);
-		}
-
 		/** @type {Record<string, any>} */
 		const props = {
 			stores: {
@@ -128,7 +123,15 @@ export async function render_response({
 				navigating: writable(null),
 				updated
 			},
-			constructors: await Promise.all(branch.map(({ node }) => node.component())),
+			constructors: await Promise.all(
+				branch.map(({ node }) => {
+					if (!node.component) {
+						// Can only be the leaf, layouts have a fallback component generated
+						throw new Error(`Missing +page.svelte component for route ${event.route.id}`);
+					}
+					return node.component();
+				})
+			),
 			form: form_value
 		};
 
@@ -321,9 +324,9 @@ export async function render_response({
 			}
 		}
 
-		// prerender a `/_app/route/path/to/page.js` module
+		// prerender a `/path/to/page/__route.js` module
 		if (manifest._.client.routes && state.prerendering && !state.prerendering.fallback) {
-			const pathname = add_resolution_prefix(event.url.pathname);
+			const pathname = add_resolution_suffix(event.url.pathname);
 
 			state.prerendering.dependencies.set(
 				pathname,