+++ packages/kit/test/apps/basics/src/hooks.js
@@ -9,7 +9,7 @@ const mapping = {
 };
 
 /** @type {import("@sveltejs/kit").Reroute} */
-export const reroute = ({ url }) => {
+export const reroute = ({ url, fetch }) => {
 	//Try to rewrite the external url used in /reroute/external to the homepage - This should not work
 	if (browser && url.href.startsWith('https://expired.badssl.com')) {
 		return '/';
@@ -25,6 +25,10 @@ export const reroute = ({ url }) => {
 
 	if (url.pathname === '/reroute/error-handling/server-error') {
 		throw new Error('Server Error - Should trigger 500 response');
+	}
+
+	if (url.pathname === '/reroute/async/a') {
+		return fetch('/reroute/api').then((r) => r.text());
 	}
 
 	if (url.pathname in mapping) {