+++ packages/kit/test/apps/basics/src/routes/reroute/async/+page.svelte
@@ -1,0 +1,1 @@
+<a href="/reroute/async/a">Go to url that should be rewritten</a>